Ui updated with new style and theme

+++ lib/views/save_screen/save_screen.dart
@@ -1,4 +1,5 @@
 import 'package:wedease/consts/consts.dart';
+import 'package:wedease/views/save_screen/save_details.dart';
 import 'package:wedease/views/service_screen/service_details.dart';
 import 'package:wedease/widgets_common/loading_indicator.dart';
 
@@ -31,19 +32,18 @@ class SaveScreen extends StatelessWidget {
             return Padding(
               padding: const EdgeInsets.all(8.0),
               child: GridView.count(
-                crossAxisCount: 2, 
-                crossAxisSpacing: 15, 
+                crossAxisCount: 2,
+                crossAxisSpacing: 15,
                 mainAxisSpacing: 20,
                 children: List.generate(data.length, (index) {
                   return GestureDetector(
                     onTap: () {
-                      // controller.checkiFav(data[index]);
                       Get.to(
-                        () => ServiceDetails(
-                            title: "${data[index]['s_name']}",
-                            data: data[index].data()),
+                        () => SaveDetails(
+                          title: "${data[index]['s_name']}",
+                          data: data[index].data(),
+                        ),
                       );
-                      print(data[index]);
                     },
                     child: Container(
                       decoration: BoxDecoration(
@@ -51,8 +51,7 @@ class SaveScreen extends StatelessWidget {
                         borderRadius: BorderRadius.circular(10),
                       ),
                       child: Column(
-                        mainAxisAlignment: MainAxisAlignment
-                            .center, 
+                        mainAxisAlignment: MainAxisAlignment.center,
                         children: [
                           Expanded(
                             child: Container(
@@ -64,7 +63,7 @@ class SaveScreen extends StatelessWidget {
                                 ),
                               ),
                               child: Image.network(
-                                "${data[index]['s_imgs']}",
+                                "${data[index]['s_imgs'][0]}",
                                 fit: BoxFit.cover,
                               ),
                             ),
@@ -83,12 +82,20 @@ class SaveScreen extends StatelessWidget {
                                           .fontFamily(semibold)
                                           .size(16)
                                           .make(),
+                                       4.heightBox,   
                                       "Rs ${data[index]['s_price']}"
                                           .text
                                           .fontFamily(semibold)
                                           .size(12)
                                           .color(borderColor)
                                           .make(),
+                                       4.heightBox,
+                                          "${data[index]['s_location']}"
+                                          .text
+                                          .fontFamily(semibold)
+                                          .size(12)
+                                          .color(borderColor)
+                                          .make(),   
                                     ],
                                   ),
                                 ),
@@ -96,8 +103,36 @@ class SaveScreen extends StatelessWidget {
                                   icon:
                                       const Icon(Icons.delete, color: redColor),
                                   onPressed: () {
-                                    FirestorServices.deleteDocument(
-                                        data[index].id);
+                                    showDialog(
+                                      context: context,
+                                      builder: (BuildContext context) {
+                                        return AlertDialog(
+                                          title:
+                                              const Text('Delete Confirmation'),
+                                          content: const Text(
+                                              'Are you sure you want to delete?'),
+                                          actions: <Widget>[
+                                            TextButton(
+                                              onPressed: () {
+                                                Navigator.of(context)
+                                                    .pop(); // Close the dialog
+                                              },
+                                              child: const Text('Cancel'),
+                                            ),
+                                            TextButton(
+                                              onPressed: () async {
+                                                FirestorServices.deleteDocument(
+                                                  data[index].id,
+                                                );
+                                                Navigator.of(context)
+                                                    .pop(); // Close the dialog
+                                              },
+                                              child: const Text('Delete'),
+                                            ),
+                                          ],
+                                        );
+                                      },
+                                    );
                                   },
                                 ),
                               ],
